backend/requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.11.16
django-annoying==0.10.4
django-anymail==4.3
django-cors-headers==2.4.0
django-extensions==2.1.3
django-tastypie==0.14.2
enum34==1.1.6
fabric==2.4.0
idna==2.7
invoke==1.2.0
ipaddress==1.0.22
kombu==4.2.1
paramiko==2.4.2
Pillow==5.3.0
psycopg2==2.7.5
pyasn1==0.4.4
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
requests==2.20.0
six==1.11.0
typing==3.6.6
urllib3==1.24
vine==1.1.4
